backend/requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
redis==5.0.3
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
PyJWT==2.8.0
